Added cache and dynamic sizing to search results

+++ methods/APISearchService.py
@@ -19,24 +19,25 @@ def GetAPISearchResultsByName(searchText, context):
         os.makedirs(helpermethods.GetProjectTempImagePath(), mode=0o777, exist_ok=True)
         with ThreadPoolExecutor(max_workers=6) as executor:
             for key, value in queryResults.items():
-                executor.submit(GetAndSaveImage, value.get('displayProperties').get('icon'), os.path.join(helpermethods.GetProjectTempImagePath(), str(key)+ ".jpg"), numOfFoundResults, context)
+                executor.submit(GetAndSaveImage, str(key), value.get('displayProperties').get('icon'), os.path.join(helpermethods.GetProjectTempImagePath(), str(key)+ ".jpg"), numOfFoundResults, context)
 
         bpy.types.WindowManager.d2ci_search_results_manager.SetQueryResultsFromSearch(queryResults)
         context.window_manager.d2ci.IsSearchingAPI = False
-
-        
         context.window_manager.d2ci.SearchResultsText = f"Found {numOfFoundResults} result{'s' if numOfFoundResults == '1' else ''}"
     except Exception as e:
         context.window_manager.d2ci.IsSearchingAPI = False
         context.window_manager.d2ci.SearchResultsText = "Failed to query the database."
 
-def GetAndSaveImage(url, downloadName, numOfFoundResults, context):
-    path = "https://bungie.net" + url
-    imageData = webcalls.Get(path)
-    if imageData.status_code == 200:
-        with open(downloadName, 'wb') as f:
-            for chunk in imageData:
-                f.write(chunk)
+def GetAndSaveImage(fileName, url, downloadName, numOfFoundResults, context):
+    if not bpy.types.WindowManager.d2ci_search_results_manager.ExistsInCache(fileName):
+        path = "https://bungie.net" + url
+        imageData = webcalls.Get(path)
+        if imageData.status_code == 200:
+            with open(downloadName, 'wb') as f:
+                for chunk in imageData:
+                    f.write(chunk)
+    else:
+        bpy.types.WindowManager.d2ci_search_results_manager.CopyFileFromCacheToTempResults(fileName)
     
     global searchResultImageCount
     searchResultImageCount += 1
